file upload possible!

+++ bpaint/admin/routes.py
@@ -1,4 +1,10 @@
-from flask import Blueprint, flash, g, redirect, render_template, request, url_for
+import os
+
+from flask import Blueprint, flash, redirect, render_template, request, url_for
+
+from random import getrandbits
+
+from werkzeug.utils import secure_filename
 
 from bpaint.admin.forms import AddToDatabaseForm, DeleteForm, UpdateForm
 
@@ -19,19 +25,20 @@ def db_add():
     form = AddToDatabaseForm()
     if request.method == 'POST':
         if form.validate_on_submit():
-            from bpaint import db
+            from bpaint import app, db, uploads
             from bpaint.models import Color
             record = form.data
-            files = []
-            for file in request.files:
-                files.append(file)
-            return str(files)
+            image = record.pop('swatch')
+            image.filename = str(getrandbits(16)) + secure_filename(image.filename)
+            with open(os.path.join(app.config['UPLOAD_FOLDER'], image.filename), 'w'):
+                image.save(os.path.join(app.config['UPLOAD_FOLDER'], image.filename))
             record.pop('csrf_token')
             record.pop('submit')
+            record['swatch'] = os.path.join(url_for('base.static', filename=f'images/{image.filename}'))
             color = Color(**record)
             db.session.add(color)
             db.session.commit()
-            flash('Success!')
+            flash(f"{record['name']} successfully added!")
             return redirect(url_for('admin.db_add'))
         else:
             return 'Error:\n' + str(form.errors)
@@ -45,26 +52,29 @@ def load_db():
 
 @bp.route('/db/update', methods=['GET', 'POST'])
 def db_update(rec_id=None):
-    if request.method == 'POST':
-        if request.form.get('update'):
-            from bpaint import db
-            from bpaint.models import Color
-            formdata = dict(request.form)
-            formdata.pop('csrf_token')
-            formdata.pop('submit')
-            record = formdata.pop('update')
-            db_rec = Color.query.filter_by(id=record).first()
-            for k, v in formdata.items():
-                setattr(db_rec, k, v)
-            db.session.add(db_rec)
-            db.session.commit()
-            flash('Update Successful!')
-            return redirect(url_for('admin.db_update'))
-        else:
-            flash('Please choose a record to update.')
-            return redirect(url_for('admin.db_update'))
-    records = load_db()
     form = UpdateForm()
+    if request.method == 'POST':
+        if form.validate_on_submit():
+            if request.form.get('update'):
+                from bpaint import db
+                from bpaint.models import Color
+                formdata = dict(request.form)
+                formdata.pop('csrf_token')
+                formdata.pop('submit')
+                record = formdata.pop('update')
+                db_rec = Color.query.filter_by(id=record).first()
+                for k, v in formdata.items():
+                    setattr(db_rec, k, v)
+                db.session.add(db_rec)
+                db.session.commit()
+                flash('Update Successful!')
+                return redirect(url_for('admin.db_update'))
+            else:
+                flash('Please choose a record to update.')
+                return redirect(url_for('admin.db_update'))
+        else:
+            return 'Error:\n' + str(form.errors)
+    records = load_db()
     form.update.choices = []
     for record in records:
         form.update.choices.append((record.id, record.name))
@@ -72,21 +82,24 @@ def db_update(rec_id=None):
 
 @bp.route('/db/delete', methods=['GET', 'POST'])
 def db_delete():
-    if request.method == 'POST':
-        if request.form.get('delete'):
-            from bpaint import db
-            from bpaint.models import Color
-            record = request.form['delete']
-            db_rec = Color.query.filter_by(id=record).first()
-            db.session.delete(db_rec)
-            db.session.commit()
-            flash('Delete Successful!')
-            return redirect(url_for('admin.db_delete'))
-        else:
-            flash('Please choose a record to delete.')
-            return redirect(url_for('admin.db_delete'))
-    records = load_db()
     form = DeleteForm()
+    if request.method == 'POST':
+        if form.validate_on_submit():
+            if request.form.get('delete'):
+                from bpaint import db
+                from bpaint.models import Color
+                record = request.form['delete']
+                db_rec = Color.query.filter_by(id=record).first()
+                db.session.delete(db_rec)
+                db.session.commit()
+                flash('Delete Successful!')
+                return redirect(url_for('admin.db_delete'))
+            else:
+                flash('Please choose a record to delete.')
+                return redirect(url_for('admin.db_delete'))
+        else:
+            return 'Error:\n' + str(form.errors)
+    records = load_db()
     form.delete.choices = []
     for record in records:
         form.delete.choices.append((record.id, record.name))
